+++ src/OpenFga.Sdk/Api/OpenFgaApi.cs
@@ -17,6 +17,7 @@ using OpenFga.Sdk.Model;
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
+using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;