+++ src/OpenFga.Sdk/Model/BatchCheckItem.cs
@@ -119,7 +119,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as BatchCheckItem);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((BatchCheckItem)input);
         }
 
         /// <summary>
@@ -152,7 +153,7 @@ namespace OpenFga.Sdk.Model {
                     (this.CorrelationId != null &&
                     this.CorrelationId.Equals(input.CorrelationId))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>