+++ src/OpenFga.Sdk/Model/BatchCheckResponse.cs
@@ -81,7 +81,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as BatchCheckResponse);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((BatchCheckResponse)input);
         }
 
         /// <summary>
@@ -100,7 +101,7 @@ namespace OpenFga.Sdk.Model {
                     input.Result != null &&
                     this.Result.SequenceEqual(input.Result)
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>