+++ src/OpenFga.Sdk/Model/ConditionMetadata.cs
@@ -90,7 +90,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as ConditionMetadata);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((ConditionMetadata)input);
         }
 
         /// <summary>
@@ -113,7 +114,7 @@ namespace OpenFga.Sdk.Model {
                     (this.SourceInfo != null &&
                     this.SourceInfo.Equals(input.SourceInfo))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>