+++ src/OpenFga.Sdk/Model/GetStoreResponse.cs
@@ -128,7 +128,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as GetStoreResponse);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((GetStoreResponse)input);
         }
 
         /// <summary>
@@ -166,7 +167,7 @@ namespace OpenFga.Sdk.Model {
                     (this.DeletedAt != null &&
                     this.DeletedAt.Equals(input.DeletedAt))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>