+++ src/OpenFga.Sdk/Model/ListUsersResponse.cs
@@ -84,7 +84,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as ListUsersResponse);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((ListUsersResponse)input);
         }
 
         /// <summary>
@@ -103,7 +104,7 @@ namespace OpenFga.Sdk.Model {
                     input.Users != null &&
                     this.Users.SequenceEqual(input.Users)
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>