+++ src/OpenFga.Sdk/Model/ReadAuthorizationModelResponse.cs
@@ -80,7 +80,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as ReadAuthorizationModelResponse);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((ReadAuthorizationModelResponse)input);
         }
 
         /// <summary>
@@ -98,7 +99,7 @@ namespace OpenFga.Sdk.Model {
                     (this.AuthorizationModel != null &&
                     this.AuthorizationModel.Equals(input.AuthorizationModel))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>