+++ src/OpenFga.Sdk/Model/RelationReference.cs
@@ -115,7 +115,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as RelationReference);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((RelationReference)input);
         }
 
         /// <summary>
@@ -148,7 +149,7 @@ namespace OpenFga.Sdk.Model {
                     (this.Condition != null &&
                     this.Condition.Equals(input.Condition))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>