+++ src/OpenFga.Sdk/Model/User.cs
@@ -100,7 +100,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as User);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((User)input);
         }
 
         /// <summary>
@@ -128,7 +129,7 @@ namespace OpenFga.Sdk.Model {
                     (this.Wildcard != null &&
                     this.Wildcard.Equals(input.Wildcard))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>