+++ src/OpenFga.Sdk/Model/Userset.cs
@@ -131,7 +131,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as Userset);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((Userset)input);
         }
 
         /// <summary>
@@ -174,7 +175,7 @@ namespace OpenFga.Sdk.Model {
                     (this.Difference != null &&
                     this.Difference.Equals(input.Difference))
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>