+++ src/OpenFga.Sdk/Model/WriteRequestWrites.cs
@@ -114,7 +114,8 @@ namespace OpenFga.Sdk.Model {
         /// <param name="input">Object to be compared</param>
         /// <returns>Boolean</returns>
         public override bool Equals(object input) {
-            return this.Equals(input as WriteRequestWrites);
+            if (input == null || input.GetType() != this.GetType()) return false;
+            return this.Equals((WriteRequestWrites)input);
         }
 
         /// <summary>
@@ -137,7 +138,7 @@ namespace OpenFga.Sdk.Model {
                     this.OnDuplicate == input.OnDuplicate ||
                     this.OnDuplicate.Equals(input.OnDuplicate)
                 )
-                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.ContainsKey(kv.Key) && Equals(kv.Value, input.AdditionalProperties[kv.Key])));
+                && (this.AdditionalProperties.Count == input.AdditionalProperties.Count && this.AdditionalProperties.All(kv => input.AdditionalProperties.TryGetValue(kv.Key, out var inputValue) && Equals(kv.Value, inputValue)));
         }
 
         /// <summary>